requirements.txt
streamlit==1.38.0
openai>=1.0.0,<2.0.0
sentence-transformers==2.7.0
scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
